+++ src/main/java/com/zerobase/zbcharger/domain/charger/dto/SearchCompanyCondition.java
@@ -1,0 +1,7 @@
+package com.zerobase.zbcharger.domain.charger.dto;
+
+public record SearchCompanyCondition(
+    Boolean deletedYn
+) {
+
+}
