--- src/main/java/com/zerobase/zbcharger/domain/charger/dto/admin/SearchCompanyRequest.java
@@ -1,7 +1,0 @@
-package com.zerobase.zbcharger.domain.charger.dto.admin;
-
-public record SearchCompanyRequest(
-    Boolean deletedYn
-) {
-
-}
